=htc small refactoring in OneHundredContinue support
